+++ examples/resources/uptime_integration_jira_servicedesk/import.sh
@@ -1,0 +1,2 @@
+# Import using the integration ID
+terraform import uptime_integration_jira_servicedesk.example 123