+++ examples/resources/uptime_integration_statuspage/import.sh
@@ -1,0 +1,2 @@
+# Import using the integration ID
+terraform import uptime_integration_statuspage.example 123