+++ examples/resources/uptime_maintenance_notification/import.sh
@@ -1,0 +1,2 @@
+# Import using the maintenance notification ID
+terraform import uptime_maintenance_notification.example 123